Adding support for serial transfers
Finally got around to doing this. SERIAL.C pretty much hooks the serial interrupts because the BIOS polling interrupts are pure trash. Thanks to this, we support up to 115200 bps (we're targetting standard 8250 hardware).

I'm not entirely happy with CTRL-BREAK behaviour. Current code can lead to lockups if CTRL-BREAK handler is invoked at certain crucial moments and we have to run it twice because I haven't figured out proper ISR chaining. But hey, at least it works and timeouts correctly if comms fail.

This implements the following changes:
 - `create_*_medium` functions now get passed the disk descriptor
 - `*_medium_ready` functions can now return `MEDIUM_READY`, `MEDIUM_NOT_READY` and `MEDIUM_RETRY`, which will repeat the last chunk transfer
 - Mediums now have an additional `*_medium_done` function that gets passed the hash data
 - The legacy disk descriptor struct now has a couple fields shuffled so that `drive_num` is on top, but the struct is still aligned (I think)